use mocha/phantomjs for tests
