+++ package.json
@@ -8,14 +8,15 @@
   },
   "scripts": {
     "build": "npm run clean && npm run build-ts && npm run lint && npm run minify && npm run cp-html",
-    "build-ts": "node_modules/.bin/tsc",
+    "build-ts": "tsc",
     "clean": "rm -rf dist",
-    "cp-html": "cp field.html button.html index.html dist",
-    "lint": "node_modules/.bin/tslint -t verbose -c tslint.json src/*.ts src/Formatter/*.ts src/Validator/*.ts src/types/*.ts src/vars/*.ts",
-    "minify": "node_modules/.bin/uglifyjs -c -m -o dist/securesubmit.min.js --source-map-root ./src  --source-map-url /dist/securesubmit.min.js.map --source-map dist/securesubmit.min.js.map --in-source-map dist/securesubmit.js.map -- dist/securesubmit.js",
+    "cp-html": "cp src/field.html src/button.html src/index.html dist",
+    "lint": "tslint -t verbose -c tslint.json src/*.ts src/Formatter/*.ts src/Validator/*.ts src/types/*.ts src/vars/*.ts",
+    "minify": "uglifyjs -c -m -o dist/securesubmit.min.js --source-map-root ./src  --source-map-url /dist/securesubmit.min.js.map --source-map dist/securesubmit.min.js.map --in-source-map dist/securesubmit.js.map -- dist/securesubmit.js",
     "install": "npm run build",
-    "test": "open 'test/index.html'",
-    "watch": "node_modules/.bin/tsc --watch"
+    "test": "phantomjs --web-security=false node_modules/mocha-phantomjs-core/mocha-phantomjs-core.js test/index.html dot",
+    "test:jsdom": "mocha test/**/*.test.js",
+    "watch": "tsc --watch"
   },
   "repository": {
     "type": "git",
@@ -37,6 +38,11 @@
   },
   "homepage": "https://github.com/hps/heartland-tokenization#readme",
   "devDependencies": {
+    "chai": "^3.5.0",
+    "jsdom": "^9.4.2",
+    "mocha": "^3.0.2",
+    "mocha-phantomjs-core": "^1.3.1",
+    "phantomjs-prebuilt": "^2.1.12",
     "tslint": "^2.5.0",
     "typescript": "^1.6.2",
     "uglify-js": "^2.4.24"
